Bug fix: the contactsMap was updating at a later time than the contactsList, which lead to out-of-sync issues in the UI in certain contexts
